Remove global install instructions
Globally installing packages is not necessary at all and can lead to weird permission issues depending on where the global install directory is. This is a friction point for new users (we just got a report on slack). So this PR rewrites that to point to the local installation instead.