Update sponsor links to inline-block
Currently, the sponsor links fill up the entire width of their parent <li>. This creates a lot of whitespace that is clickable, even though it can be unclear what is being clicked on (the image does change color to provide a hint, though). This change prevents the <a> tag from filling up the entire space of the <li>. So now, only the image is clickable and not the whitespace around it.