docs: TS switch from `type` to `interfaces`
These days the TypeScript recommends to use the `interface` keyword instead of `type` in their docs:

https://www.typescriptlang.org/docs/handbook/typescript-in-5-minutes.html

> You’ll see that there are two syntaxes for building types: Interfaces and Types. You should prefer interface. Use type when you need specific features.

For typing component props whether you use `type` or `interface` hardly matters, since the differences mainly show up when you combine or extend types which is rare with props. That said it's better for us to align with the TS docs.